requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
asyncio==3.4.3
aiofiles==23.2.1
aiohttp==3.9.1
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
pillow==10.1.0
opencv-python==4.8.1.78
qrcode==7.4.2
cryptography==41.0.8
schedule==1.2.0
apscheduler==3.10.4
pydantic==2.5.2
bcrypt==4.1.2
jwt==1.3.1
pyjwt==2.8.0
requests==2.31.0